usermanagement finished

+++ server/routes/manageUser.js
@@ -8,6 +8,12 @@ const router = express.Router();
 const log = console.log;
 
 router.post('/assignAdmin', (req, res) => {
+    if (mongoose.connection.readyState != 1) {
+        log('Issue with mongoose connection')
+        res.status(500).send('Internal server error')
+        return;
+    }
+
     User.findOne({ email: req.body.email }).then((user) => {
         if (!user) {
             res.status(404).send('user not found')
@@ -29,6 +35,11 @@ router.post('/assignAdmin', (req, res) => {
 })
 
 router.post('/deleteUser', (req, res) => {
+    if (mongoose.connection.readyState != 1) {
+        log('Issue with mongoose connection')
+        res.status(500).send('Internal server error')
+        return;
+    }
 
     User.findOneAndDelete({ email: req.body.email }).then((user) => {
         if (!user) {
@@ -41,6 +52,16 @@ router.post('/deleteUser', (req, res) => {
         log(error)
         res.status(500).send('Internal Server Error')
     })
+})
+
+router.get('/getUsers', (req, res) => {
+    User.find().then(users => {
+        res.send({ users })
+    }).catch((error) => {
+        log(error)
+        res.status(500).send('Internal Server Error')
+    })
+
 })
